requirements.txt
altair
annotated-doc
annotated-types
anyio
attrs
blinker
cachetools
certifi
charset-normalizer
click
colorama
dnspython
email-validator
fastapi
fastapi-cli
fastapi-cloud-cli
fastar
gitdb
GitPython
h11
httpcore
httptools
httpx
idna
Jinja2
jsonschema
jsonschema-specifications
markdown-it-py
MarkupSafe
mdurl
narwhals
numpy
packaging
pandas
pillow
protobuf
pyarrow
pydantic
pydantic-extra-types
pydantic-settings
pydantic_core
pydeck
Pygments
python-dateutil
python-dotenv
python-multipart
pytz
PyYAML
referencing
requests
rich
rich-toolkit
rignore
rpds-py
sentry-sdk
shellingham
six
smmap
starlette
streamlit
streamlit-float
tenacity
toml
tornado
typer
typing-inspection
typing_extensions
tzdata
urllib3
uvicorn
watchdog
watchfiles
websockets
